# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Book Any Car
  description: Book Any Car apis docs
  version: 1.0.0

servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://bookanycar.herokuapp.com/api/v1/
    description: Production Server

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer




paths:
  /cars:
    get:
      summary: get all plans
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          description: number of page(start from 0)
          required: false
        - name: size
          in: query
          schema:
            type: integer
          description: number of plans on one page
          required: false
        - name: carname
          in: query
          schema:
            type: string
          description: filter by carname
          required: false
      responses:
        '200':
          description: A json array of plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPlans:
                    type: number 
                  plans:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        car:
                          type: string
                        model:
                          type: string
                        fair:
                          type: string
                        img:
                          type: string
  /car:
    post:
      summary: post a new car
      requestBody:
        description: adding a new car
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                car:
                  type: string
                model:
                  type: string
                fair:
                  type: string
                img:
                  type: string
                  format: binary
      responses:
        '200':
          description: A json array of plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  car:
                    type: string
                  model:
                    type: string
                  fair:
                    type: string
                  img:
                    type: string
  /user/signup:
    post: 
      summary: Sign up a new user
      requestBody:
        description: adding a new user
        content:
          application/json:
            schema:
              type: object
              properties:
                user: 
                  type: string
                  example: rahul123
                password:
                  type: string
                  example: rahul@123
                email:
                  type: string
                  example: example@example.com
                role:
                  type: string 
                  default: user
                  example: user     
      responses:
        200: 
          description: A json object of user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  user:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
  /user/login:
    post:
      summary: login user
      requestBody:
        description: adding a new user
        content:
          application/json:
            schema:
              type: object
              properties:
                user: 
                  type: string
                  example: rahul123
                password:
                  type: string
                  example: rahul@123
                email:
                  type: string
                  example: example@example.com
      responses:
        200: 
          description: A json object of user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
      
          